requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
altair==5.4.0
annotated-types==0.7.0
asgiref==3.8.1
attrs==24.2.0
backoff==2.2.1
bcrypt==4.2.0
blinker==1.8.2
build==1.2.1
cachetools==5.5.0
certifi==2024.7.4
chroma-hnswlib==0.7.6
chromadb==0.5.5
click==8.1.7
coloredlogs==15.0.1
dataclasses-json==0.6.7
Deprecated==1.2.14
fastapi==0.112.2
flatbuffers==24.3.25
frozenlist==1.4.1
fsspec==2024.6.1
gitdb==4.0.11
GitPython==3.1.43
google-auth==2.34.0
googleapis-common-protos==1.65.0
greenlet==3.0.3
grpcio==1.66.0
httptools==0.6.1
huggingface-hub==0.24.6
humanfriendly==10.0
idna==3.8
importlib_resources==6.4.4
joblib==1.4.2
jsonpatch==1.33
jsonschema==4.23.0
kubernetes==30.1.0
langchain==0.2.15
langchain-community==0.2.13
langchain-core==0.2.35
langchain-huggingface==0.0.3
langchain-text-splitters==0.2.2
langsmith==0.1.106
marshmallow==3.22.0
mmh3==4.1.0
monotonic==1.6
mpmath==1.3.0
multidict==6.0.5
narwhals==1.5.5
networkx==3.3
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.20
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
onnxruntime==1.19.0
opentelemetry-api==1.26.0
opentelemetry-exporter-otlp-proto-common==1.26.0
opentelemetry-exporter-otlp-proto-grpc==1.26.0
opentelemetry-instrumentation==0.47b0
opentelemetry-instrumentation-asgi==0.47b0
opentelemetry-instrumentation-fastapi==0.47b0
opentelemetry-proto==1.26.0
opentelemetry-sdk==1.26.0
opentelemetry-semantic-conventions==0.47b0
opentelemetry-util-http==0.47b0
orjson==3.10.7
pandas==2.2.2
pillow==10.4.0
posthog==3.6.0
protobuf==4.25.4
pyarrow==17.0.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydantic==2.8.2
pydantic_core==2.20.1
pydeck==0.9.1
pypdf==4.3.1
PyPika==0.48.9
pyproject_hooks==1.1.0
python-dotenv==1.0.1
pytz==2024.1
regex==2024.7.24
requests-oauthlib==2.0.0
rich==13.8.0
rpds-py==0.20.0
rsa==4.9
safetensors==0.4.4
scikit-learn==1.5.1
scipy==1.14.1
sentence-transformers==3.0.1
shellingham==1.5.4
smmap==5.0.1
SQLAlchemy==2.0.32
starlette==0.38.2
streamlit==1.37.1
swig==4.2.1
sympy==1.13.2
tenacity==8.5.0
threadpoolctl==3.5.0
tokenizers==0.19.1
toml==0.10.2
torch==2.4.0
tqdm==4.66.5
transformers==4.44.2
triton==3.0.0
typer==0.12.5
typing-inspect==0.9.0
tzdata==2024.1
uvicorn==0.30.6
uvloop==0.20.0
watchdog==4.0.2
watchfiles==0.23.0
websockets==13.0
wrapt==1.16.0
yarl==1.9.4



